requirements.txt
pandas
openpyxl
pdfplumber
anthropic
streamlit
python-dotenv
numpy
Pillow
